DCES-552 Refactor client configuration
