+++ dces-drc-integration/src/main/java/uk/gov/justice/laa/crime/dces/integration/config/DrcApiWebClientConfiguration.java
@@ -17,7 +17,6 @@ import org.springframework.web.service.invoker.HttpServiceProxyFactory;
 import reactor.netty.http.client.HttpClient;
 import reactor.netty.resources.ConnectionProvider;
 import uk.gov.justice.laa.crime.dces.integration.client.DrcClient;
-import uk.gov.justice.laa.crime.dces.integration.maatapi.config.ServicesConfiguration;
 
 import javax.net.ssl.SSLException;
 import java.time.Duration;
