--- dces-drc-integration/src/main/java/uk/gov/justice/laa/crime/dces/integration/maatapi/client/MaatApiClient.java
@@ -1,5 +1,0 @@
-package uk.gov.justice.laa.crime.dces.integration.maatapi.client;
-
-public interface MaatApiClient {
-
-}
