+++ dces-drc-integration/src/test/java/uk/gov/justice/laa/crime/dces/integration/maatapi/MaatApiClientFactoryTest.java
@@ -5,7 +5,8 @@ import org.junit.jupiter.api.TestInstance;
 import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.boot.test.mock.mockito.MockBean;
 import org.springframework.web.reactive.function.client.WebClient;
-import uk.gov.justice.laa.crime.dces.integration.maatapi.client.MaatApiClient;
+import uk.gov.justice.laa.crime.dces.integration.client.MaatApiClient;
+import uk.gov.justice.laa.crime.dces.integration.config.MaatApiClientFactory;
 
 import static org.assertj.core.api.AssertionsForClassTypes.assertThat;
 @SpringBootTest
